+++ src/indication/indication.cpp
@@ -8,12 +8,12 @@
 // ---------------------------------
 #define LED_BLINK_TIME_MS 50
 
-static const uint8_t LED_PINS[] = {
+static const uint16_t LED_PINS[] = {
     LED_LIGHT_MANUAL_PIN, LED_LIGHT_AUTO_PIN,      LED_LIGHT_MIN_PIN,
     LED_LIGHT_MAX_PIN,    LED_TEMPERATURE_MIN_PIN, LED_TEMPERATURE_MAX_PIN,
     LED_HUMIDITY_MIN_PIN, LED_SILENT_PIN};
 
-static const uint8_t LED_MASKS[] = {
+static const uint16_t LED_MASKS[] = {
     LED_LIGHT_COMMAND_MASK, LED_LIGHT_AUTO_MASK,      LED_LIGHT_MIN_MASK,
     LED_LIGHT_MAX_MASK,     LED_TEMPERATURE_MIN_MASK, LED_TEMPERATURE_MAX_MASK,
     LED_HUMIDITY_MIN_MASK,  LED_SILENT_MASK};
@@ -32,8 +32,8 @@ void initIndication() {
   pinMode(LED_SILENT_PIN, OUTPUT);
 }
 
-void handleIndication(uint8_t &ledState) {
-  for (uint8_t i = 0; i < sizeof(LED_PINS) / sizeof(LED_PINS[0]); ++i) {
+void handleIndication(uint16_t &ledState) {
+  for (uint16_t i = 0; i < sizeof(LED_PINS) / sizeof(LED_PINS[0]); ++i) {
     digitalWrite(LED_PINS[i], (ledState & LED_MASKS[i]) != 0);
   }
 }
